+++ R/Sessions.R
@@ -72,12 +72,9 @@ api_get_invite <- function(url, api_key, app_uid, user_uid) {
 Sessions <-  R6::R6Class(
   classname = 'Sessions',
   public = list(
-    hosted_url = character(0),
-    app_name = character(0),
     conn = NULL,
     firebase_config = NULL,
     is_invite_required = TRUE,
-    api_key = NULL,
     sign_in_providers = character(0),
     is_email_verification_required = TRUE,
     #' @description
@@ -90,31 +87,17 @@ Sessions <-  R6::R6Class(
     #' @inheritParams global_sessions_config
     #'
     config = function(
-      app_name,
-      api_key,
       firebase_config = NULL,
       admin_mode = FALSE,
       is_invite_required = TRUE,
-      api_url = "https://api.polished.tech",
       sign_in_providers = "email",
       is_email_verification_required = FALSE
     ) {
-
-      if (!(length(app_name) == 1 && is.character(app_name))) {
-        stop("invalid `app_name` argument passed to `global_sessions_config()`", call. = FALSE)
-      }
-
-      if (!(length(api_url) == 1 && is.character(api_url))) {
-        stop("invalid `api_url` argument passed to `global_sessions_config()`", call. = FALSE)
-      }
 
       if (!(length(sign_in_providers) >= 1 && is.character(sign_in_providers))) {
         stop("invalid `sign_in_providers` argument passed to `global_sessions_config()`", call. = FALSE)
       }
 
-
-      self$api_key <- api_key
-      self$hosted_url <- api_url
       self$sign_in_providers <- sign_in_providers
 
       if (!is.null(firebase_config)) {
@@ -124,34 +107,6 @@ Sessions <-  R6::R6Class(
         }
         self$firebase_config <- firebase_config
       }
-
-
-
-      # get the app uid
-      res <- httr::GET(
-        url = paste0(self$hosted_url, "/apps"),
-        query = list(
-          app_name = app_name
-        ),
-        httr::authenticate(
-          user = self$api_key,
-          password = ""
-        )
-      )
-
-      app <- jsonlite::fromJSON(
-        httr::content(res, "text", encoding = "UTF-8")
-      )
-
-      if (!identical(httr::status_code(res), 200L)) {
-        stop(app, call. = FALSE)
-      }
-
-      if (length(app) == 0) {
-        stop(paste0("app_name `", app_name, "` does not exist"), call. = FALSE)
-      }
-
-      self$app_name <- app$uid
 
 
 
@@ -222,25 +177,25 @@ Sessions <-  R6::R6Class(
 
 
         invite <- api_get_invite_by_email(
-          self$hosted_url,
-          self$api_key,
+          getOption("polished")$api_url,
+          getOption("polished")$api_key,
           new_session$email,
-          self$app_name
+          getOption("polished")$app_uid
         )
 
         if (isFALSE(self$is_invite_required) && is.null(invite)) {
           # if invite is not required, and this is the first time that the user is signing in,
           # then create the app_users
           res <- httr::POST(
-            url = paste0(self$hosted_url, "/app-users"),
+            url = paste0(getOption("polished")$api_url, "/app-users"),
             body = list(
               email = new_session$email,
-              app_uid = self$app_name,
+              app_uid = getOption("polished")$app_uid,
               is_admin = FALSE,
               req_user_uid = "00000000-0000-0000-0000-000000000000"
             ),
             httr::authenticate(
-              user = self$api_key,
+              user = getOption("polished")$api_key,
               password = ""
             ),
             encode = "json"
@@ -249,10 +204,10 @@ Sessions <-  R6::R6Class(
           httr::stop_for_status(res)
 
           invite <- api_get_invite_by_email(
-            self$hosted_url,
-            self$api_key,
+            getOption("polished")$api_url,
+            getOption("polished")$api_key,
             new_session$email,
-            self$app_name
+            getOption("polished")$app_uid
           )
 
         }
@@ -279,10 +234,10 @@ Sessions <-  R6::R6Class(
       invite <- NULL
 
       invite <- api_get_invite_by_email(
-        self$hosted_url,
-        self$api_key,
+        getOption("polished")$api_url,
+        getOption("polished")$api_key,
         email,
-        self$app_name
+        getOption("polished")$app_uid
       )
 
       return(invite)
@@ -291,14 +246,14 @@ Sessions <-  R6::R6Class(
       if (nchar(hashed_cookie) == 0) return(NULL)
 
       res <- httr::GET(
-        url = paste0(self$hosted_url, "/session-by-cookie"),
+        url = paste0(getOption("polished")$api_url, "/session-by-cookie"),
         query = list(
           hashed_cookie = hashed_cookie,
-          app_uid = self$app_name,
+          app_uid = getOption("polished")$app_uid,
           page = page
         ),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         encode = "json"
@@ -320,9 +275,9 @@ Sessions <-  R6::R6Class(
     sign_in_email = function(email, password, hashed_cookie) {
 
       res <- httr::POST(
-        url = paste0(self$hosted_url, "/sign-in-email"),
+        url = paste0(getOption("polished")$api_url, "/sign-in-email"),
         body = list(
-          app_uid = self$app_name,
+          app_uid = getOption("polished")$app_uid,
           email = email,
           password = password,
           hashed_cookie = hashed_cookie,
@@ -330,7 +285,7 @@ Sessions <-  R6::R6Class(
         ),
         encode = "json",
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         )
       )
@@ -347,14 +302,14 @@ Sessions <-  R6::R6Class(
 
           # send a password reset email and stop
           res2 <- httr::POST(
-            url = paste0(self$hosted_url, "/send-password-reset-email"),
+            url = paste0(getOption("polished")$api_url, "/send-password-reset-email"),
             body = list(
               email = email,
-              app_uid = self$app_name,
+              app_uid = getOption("polished")$app_uid,
               is_invite_required = self$is_invite_required
             ),
             httr::authenticate(
-              user = self$api_key,
+              user = getOption("polished")$api_key,
               password = ""
             ),
             encode = "json"
@@ -382,13 +337,13 @@ Sessions <-  R6::R6Class(
     register_email = function(email, password, hashed_cookie) {
 
       res <- httr::POST(
-        url = paste0(self$hosted_url, "/register-email"),
+        url = paste0(getOption("polished")$api_url, "/register-email"),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
-          app_uid = self$app_name,
+          app_uid = getOption("polished")$app_uid,
           email = email,
           password = password,
           hashed_cookie = hashed_cookie,
@@ -436,9 +391,9 @@ Sessions <-  R6::R6Class(
       } else {
 
         res <- httr::PUT(
-          url = paste0(self$hosted_url, "/sessions"),
+          url = paste0(getOption("polished")$api_url, "/sessions"),
           httr::authenticate(
-            user = self$api_key,
+            user = getOption("polished")$api_key,
             password = ""
           ),
           body = list(
@@ -459,9 +414,9 @@ Sessions <-  R6::R6Class(
     set_signed_in_as = function(session_uid, signed_in_as, user_uid = NULL) {
 
       res <- httr::PUT(
-        url = paste0(self$hosted_url, "/sessions"),
+        url = paste0(getOption("polished")$api_url, "/sessions"),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
@@ -481,21 +436,21 @@ Sessions <-  R6::R6Class(
     get_signed_in_as_user = function(user_uid) {
 
       invite <- api_get_invite(
-        self$hosted_url,
-        self$api_key,
-        self$app_name,
+        getOption("polished")$api_url,
+        getOption("polished")$api_key,
+        getOption("polished")$app_uid,
         user_uid
       )
 
       email <- invite$email
 
       res <- httr::GET(
-        url = paste0(self$hosted_url, "/user-roles"),
+        url = paste0(getOption("polished")$api_url, "/user-roles"),
         query = list(
           user_uid = user_uid
         ),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         encode = "json"
@@ -526,9 +481,9 @@ Sessions <-  R6::R6Class(
 
 
       res <- httr::POST(
-        url = paste0(self$hosted_url, "/actions"),
+        url = paste0(getOption("polished")$api_url, "/actions"),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
@@ -544,9 +499,9 @@ Sessions <-  R6::R6Class(
     sign_out = function(hashed_cookie, session_uid) {
 
       res <- httr::POST(
-        url = paste0(self$hosted_url, "/sign-out"),
+        url = paste0(getOption("polished")$api_url, "/sign-out"),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
@@ -567,14 +522,14 @@ Sessions <-  R6::R6Class(
 
       # add session to "sessions" table via the API
       res <- httr::POST(
-        url = paste0(self$hosted_url, "/sessions"),
+        url = paste0(getOption("polished")$api_url, "/sessions"),
         httr::authenticate(
-          user = self$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
           data = session_data,
-          app_uid = self$app_name
+          app_uid = getOption("polished")$app_uid
         ),
         encode = "json"
       )