+++ R/verify_email_module.R
@@ -87,9 +87,9 @@ verify_email_module <- function(input, output, session) {
         tryCatch({
 
           res <- httr::GET(
-            url = paste0(.global_sessions$hosted_url, "/users"),
+            url = paste0(getOption("polished")$api_url, "/users"),
             httr::authenticate(
-              user = .global_sessions$api_key,
+              user = getOption("polished")$api_key,
               password = ""
             ),
             query = list(
@@ -134,15 +134,15 @@ verify_email_module <- function(input, output, session) {
 
 
       res <- httr::POST(
-        url = paste0(.global_sessions$hosted_url, "/resend-verification-email"),
+        url = paste0(getOption("polished")$api_url, "/resend-verification-email"),
         httr::authenticate(
-          user = .global_sessions$api_key,
+          user = getOption("polished")$api_key,
           password = ""
         ),
         body = list(
           email = hold_email,
           user_uid = session$userData$user()$user_uid,
-          app_uid = .global_sessions$app_name
+          app_uid = getOption("polished")$app_uid
         ),
         encode = "json"
       )